requirements.txt
jupyterlab==3.6.3
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.24.3
pandas==2.0.1
scipy==1.10.1
statsmodels==0.13.5
scikit-learn==1.2.2
ipympl==0.9.3
gTTS==2.4.0
moviepy==1.0.3
SpeechRecognition
python-levenshtein
streamlit
streamlit-audiorec
st-annotated-text
portaudio
pyaudio
os
pronouncing
difflib
playsound
IPython
